+++ src/components/dashboard/Dashboard.js
@@ -1,32 +1,20 @@
 import { lazy, Suspense, useCallback, useEffect } from 'react'
 import { useState } from 'react'
-import { getAttendences, getBirthdates, getWeather, getDisplayMode, getOS, newFeedback, updateAttendences, getEvalByEvent } from '../../modules/data/DBConnect'
+import { getAttendences, getBirthdates, getDisplayMode, getOS, newFeedback } from '../../modules/data/DBConnect'
 import { StyledChangelog, StyledDashboard, StyledFeedbackArea, StyledInfoText } from './Dashboard.styled'
-import { Clothing } from '../../modules/components/icons/Clothing'
 import { TbAlertTriangle } from 'react-icons/tb'
 import { IoShareOutline } from 'react-icons/io5'
 import { BsPlusSquare } from 'react-icons/bs'
 import { beforeInstallPrompt } from '../..'
 
-import {
-    Chart as ChartJS,
-    CategoryScale,
-    LinearScale,
-    BarElement,
-    Title,
-    Tooltip,
-    Legend,
-  } from 'chart.js'
-import { Bar } from "react-chartjs-2"
 import { version } from '../../App'
-import EventInfo from './EventInfo'
-import Statistics from './Statistics'
-import PlusOne from '../../modules/components/icons/PlusOne'
-import { FaUserGroup } from 'react-icons/fa6'
+import EventInfo from './eventinfo/EventInfo'
+import Statistics from './statistics/Statistics'
+import NextEvent from './events/NextEvent'
+import NextPractice from './events/NextPractice'
+import NextOther from './events/NextOther'
 
 const Button = lazy(() => import('../../modules/components/button/Button'))
-const Terminzusage = lazy(() => import('../dateplanner/attendenceInput/Terminzusage'))
-const WeatherIcon = lazy(() => import('./WeatherIcon'))
 
 const Dashboard = ({ fullname, auth_level, theme }) => {
 
@@ -249,271 +237,6 @@ const Changelog = ({read}) => {
     )
 }
 
-const ClothingData = ({ clothing, onClick }) => {
-
-    return(
-        <Suspense>
-            {parseInt(clothing) !== 0 ? <>
-                <td onClick={onClick}>Bekleidung:</td>
-                <td onClick={onClick}><Clothing clothing={parseInt(clothing)} /></td>
-            </> : <><td colSpan={2}></td></>}
-        </Suspense>
-    )
-}
-
-const PlusOneData = ({ attendence, plusOne, callback, theme }) => {
-
-    const onClick = () => {
-        if(attendence === 1)
-            callback()
-    }
-
-    return(<td><PlusOne plusOne={plusOne} active={attendence === 1} onClick={onClick} theme={theme} /></td>)
-}
-
-const NextPractice = ({ nextPractice, auth_level, showEventInfo, theme }) => {
-
-    const [evaluation, setEvaluation] = useState()
-
-    let practiceDate = new Date(nextPractice?.Date)
-    let attendence = nextPractice?.Attendence
-
-    const onClick = async (event_id, att) => {
-        let changes = {}
-        changes['' + event_id] = [att, false]
-        await updateAttendences(changes, false)
-        updateEventEval()
-    }
-
-    const updateEventEval = useCallback(async () => {
-        let _eval = await getEvalByEvent(nextPractice?.Event_ID, nextPractice?.Usergroup_ID)
-        setEvaluation(_eval)
-        return
-    }, [nextPractice])
-
-    const clickTD = useCallback(() => {
-        showEventInfo(nextPractice)
-    }, [showEventInfo, nextPractice])
-
-    useEffect(() => {
-        if(nextPractice !== undefined){
-            updateEventEval()
-        }
-    }, [nextPractice, updateEventEval])
-
-    useEffect(() => {
-        const interval = setInterval(() => {
-            updateEventEval()
-        }, 60000);
-        return () => clearInterval(interval);
-      }, [updateEventEval]);
-
-    return(<>
-        <tr className='event_header'>
-            <td onClick={clickTD}>{nextPractice?.Type}</td>
-            <td onClick={clickTD}>{nextPractice?.Location}</td>
-            <td rowSpan={2}><Terminzusage event={nextPractice} event_id={nextPractice?.Event_ID} states={3} attendence={attendence} onClick={onClick} cancelled={nextPractice?.Type.includes('Abgesagt')} theme={theme}/></td>
-        </tr>
-        <tr>
-            <td onClick={clickTD}>{practiceDate.getDate()}.{practiceDate.getMonth() + 1}.{practiceDate.getFullYear()}</td>
-            <td onClick={clickTD}>{nextPractice?.Begin === null ? '-' : `${nextPractice?.Begin.slice(0, 5)} Uhr`}</td>
-        </tr>
-        <tr>
-            {auth_level > 1 ? <td colSpan={3}><DashboardDiagram event={evaluation} auth_level={auth_level} theme={theme}/></td> : <></>}
-        </tr>
-    </>)
-}
-
-const NextOther = ({ nextOther, auth_level, showEventInfo, theme }) => {
-
-    const [weather, setWeather] = useState()
-    const [evaluation, setEvaluation] = useState()
-    const [attendence, setAttendence] = useState(nextOther?.Attendence)
-    const [plusone, setPlusOne] = useState(nextOther?.PlusOne)
-
-    let gigDate = new Date(nextOther?.Date)
-
-    const onClick = async (event_id, att) => {
-        let changes = {}
-        changes['' + event_id] = [att, plusone]
-        setAttendence(att)
-        await updateAttendences(changes, false)
-        updateEventEval()
-    }
-
-    const updatePlusOne = async () => {
-        let changes = {}
-        changes['' + nextOther?.Event_ID] = [attendence, !plusone]
-        await updateAttendences(changes, false)
-        setPlusOne(!plusone)
-        updateEventEval()
-    }
-
-    const updateEventEval = useCallback(async () => {
-        let _eval = await getEvalByEvent(nextOther?.Event_ID, nextOther?.Usergroup_ID)
-        setEvaluation(_eval)
-        return
-    }, [nextOther])
-
-    const clickTD = useCallback(() => {
-        showEventInfo(nextOther)
-    }, [showEventInfo, nextOther])
-
-    useEffect(() => {
-        let eDate = new Date(nextOther?.Date)
-        let nextWeek = new Date()
-        nextWeek.setDate(nextWeek.getDate() + 6)
-        if(nextOther !== undefined && eDate < nextWeek) {
-            getWeather(nextOther).then(weather => {
-                setWeather(weather)
-            })
-        }
-        if(nextOther !== undefined){
-            updateEventEval()
-        }
-    }, [nextOther, updateEventEval])
-
-    useEffect(() => {
-        const interval = setInterval(() => {
-            updateEventEval()
-        }, 60000);
-        return () => clearInterval(interval);
-    }, [updateEventEval]);
-    
-    return(<>
-        <tr className='event_header'>
-            <td onClick={clickTD}>{nextOther?.Type}</td>
-            <td onClick={clickTD}>{nextOther?.Location}</td>
-        </tr>
-        <tr>
-            <td onClick={clickTD}>{gigDate.getDate()}.{gigDate.getMonth() + 1}.{gigDate.getFullYear()}</td>
-            <td onClick={clickTD}>{nextOther?.Begin !== "12:34:56" && nextOther?.Begin !== null ? `${nextOther?.Begin.slice(0, 5)} Uhr` : "-"}</td>
-            <td rowSpan={3}><Suspense><Terminzusage event={nextOther} event_id={nextOther?.Event_ID} states={3} attendence={attendence} onClick={onClick} cancelled={nextOther?.Type.includes('Abgesagt')} theme={theme}/></Suspense></td>
-        </tr>
-        <tr>
-            <td onClick={clickTD}>Hin:</td>
-            <td onClick={clickTD}>{nextOther?.Departure !== "12:34:56" && nextOther?.Departure !== null ? `${nextOther?.Departure.slice(0, 5)} Uhr` : "-"}</td>
-        </tr>
-        <tr>
-            <td onClick={clickTD}>Zurück:</td>
-            <td onClick={clickTD}>{nextOther?.Leave_dep !== "12:34:56" && nextOther?.Leave_dep !== null ? `${nextOther?.Leave_dep.slice(0, 5)} Uhr` : "-"}</td>
-        </tr>
-        <tr>
-            <ClothingData  onClick={clickTD} clothing={nextOther?.Clothing} />
-            {nextOther?.Ev_PlusOne ? <PlusOneData attendence={attendence} theme={theme} callback={updatePlusOne} plusOne={plusone} /> : <></>}            
-        </tr>
-        {weather ? <Suspense>
-            <tr>
-                <td>Wetter:</td>
-                <td>{`${weather.Temperature}°C`}</td>
-                <td><WeatherIcon code={weather.Weathercode} /></td>
-            </tr>
-        </Suspense> : <></>}
-        <tr>
-            {auth_level > 0 ? <td colSpan={3}><DashboardDiagram event={evaluation} auth_level={auth_level} theme={theme}/></td> : <></>}
-        </tr>
-        {nextOther?.Ev_PlusOne && evaluation?.PlusOne > 0 ? <tr>
-            <td><FaUserGroup className='Plusone_icon'/></td>
-            <td>+{evaluation?.PlusOne}</td>
-        </tr> : <></>}
-    </>)
-}
-
-const NextEvent = ({ nextEvent, auth_level, showEventInfo, theme }) => {
-
-    const [weather, setWeather] = useState()
-    const [evaluation, setEvaluation] = useState()
-    const [attendence, setAttendence] = useState(nextEvent?.Attendence)
-    const [plusone, setPlusOne] = useState(nextEvent?.PlusOne)
-    let eventDate = new Date(nextEvent?.Date)
-
-    const onClick = async (event_id, att) => {
-        let changes = {}
-        changes['' + event_id] = [att, plusone]
-        await updateAttendences(changes, false)
-        setAttendence(att)
-        updateEventEval()
-    }
-
-    const updatePlusOne = async () => {
-        let changes = {}
-        changes['' + nextEvent?.Event_ID] = [attendence, !plusone]
-        await updateAttendences(changes, false)
-        setPlusOne(!plusone)
-        updateEventEval()
-    }
-
-    const updateEventEval = useCallback(async () => {
-        let _eval = await getEvalByEvent(nextEvent?.Event_ID, nextEvent?.Usergroup_ID)
-        setEvaluation(_eval)
-        return
-    }, [nextEvent])
-
-    const clickTD = useCallback(() => {
-        showEventInfo(nextEvent)
-    }, [showEventInfo, nextEvent])
-
-    useEffect(() => {
-        let eDate = new Date(nextEvent?.Date)
-        let nextWeek = new Date()
-        nextWeek.setDate(nextWeek.getDate() + 6)
-        if(nextEvent !== undefined && eDate < nextWeek) {
-            getWeather(nextEvent).then(weather => {
-                setWeather(weather)
-            })
-        }
-        if(nextEvent !== undefined){
-            updateEventEval()
-        }
-    }, [nextEvent, updateEventEval])
-
-    useEffect(() => {
-        const interval = setInterval(() => {
-            updateEventEval()
-        }, 60000);
-        return () => clearInterval(interval);
-      }, [updateEventEval]);
-
-    return(<>
-        <tr className='event_header'>
-            <td onClick={clickTD}>{nextEvent?.Type}</td>
-            <td onClick={clickTD}>{nextEvent?.Location}</td>
-        </tr>
-        <tr>
-            <td onClick={clickTD}>{eventDate.getDate()}.{eventDate.getMonth() + 1}.{eventDate.getFullYear()}</td>
-            <td onClick={clickTD}>{nextEvent?.Begin !== "12:34:56" && nextEvent?.Begin !== null ? `${nextEvent?.Begin.slice(0, 5)} Uhr` : "-"}</td>
-            <td rowSpan={3}><Suspense><Terminzusage event={nextEvent} event_id={nextEvent?.Event_ID} states={3} attendence={attendence} onClick={onClick} cancelled={nextEvent?.Type.includes('Abgesagt')} theme={theme}/></Suspense></td>
-        </tr>
-        <tr>
-            <td onClick={clickTD}>Hin:</td>
-            <td onClick={clickTD}>{nextEvent?.Departure !== "12:34:56" && nextEvent?.Departure !== null ? `${nextEvent?.Departure.slice(0, 5)} Uhr` : "-"}</td>
-        </tr>
-        <tr>
-            <td onClick={clickTD}>Zurück:</td>
-            <td onClick={clickTD}>{nextEvent?.Leave_dep !== "12:34:56" && nextEvent?.Leave_dep !== null ? `${nextEvent?.Leave_dep.slice(0, 5)} Uhr` : "-"}</td>
-        </tr>
-        <tr>
-            <ClothingData  onClick={clickTD} clothing={nextEvent?.Clothing} />
-            {nextEvent?.Ev_PlusOne ? <PlusOneData attendence={attendence} theme={theme} callback={updatePlusOne} plusOne={plusone}/> : <></>}
-        </tr>
-        {weather ? <Suspense>
-            <tr>
-                <td>Wetter:</td>
-                <td>{`${weather.Temperature}°C`}</td>
-                <td><WeatherIcon code={weather.Weathercode} /></td>
-            </tr>
-        </Suspense> : <></>}
-        <tr>
-            {auth_level > 0 ? <td colSpan={3}><DashboardDiagram event={evaluation} auth_level={auth_level} theme={theme}/></td> : <></>}
-        </tr>
-        {nextEvent?.Ev_PlusOne && evaluation?.PlusOne > 0 ? <tr>
-            <td><FaUserGroup className='Plusone_icon'/></td>
-            <td>+{evaluation?.PlusOne}</td>
-        </tr> : <></>}
-        
-    </>)
-}
-
 const Feedback = () => {
 
     const [open, setOpen] = useState(false)
@@ -543,70 +266,6 @@ const Feedback = () => {
     </div>)
 }
 
-const DashboardDiagram = ({ event, auth_level, theme }) => {
 
-    ChartJS.register(
-        CategoryScale,
-        LinearScale,
-        BarElement,
-        Title,
-        Tooltip,
-        Legend
-    )
-
-    const options = {
-        animation: {
-            duration: 0
-        },
-        indexAxis: 'y',
-        plugins: {
-            title: {
-                display: false
-            },
-            legend: {
-                display: false
-            },
-            tooltip: {
-                enabled: auth_level > 1
-            }
-        },
-        scales: {
-            x: {
-                stacked: true,
-                display: false
-            },
-            y: {
-                stacked: true,
-                display: false
-            }
-        }
-    }
-
-    const labels = ['']
-
-    const data = {
-        labels,
-        datasets: [
-            {
-                data: [event?.Consent],
-                backgroundColor: theme.greenRGB
-            },
-            {
-                data: [event?.Missing],
-                backgroundColor: theme.blueRGB
-            },
-            {
-                data: [event?.Maybe],
-                backgroundColor: theme.yellowRGB
-            },
-            {
-                data: [event?.Refusal],
-                backgroundColor: theme.redRGB
-            }
-        ]
-    }
-
-    return(<Bar height={"30px"} options={options} data={data}/>)
-}
 
 export default Dashboard
