+++ src/components/dashboard/eventinfo/EventInfo.styled.js
